requirements.txt
cyclonedx-python-lib
jsonschema